+++ builder/rollup/rollup.game.mjs
@@ -6,14 +6,18 @@ import postcss from 'rollup-plugin-postcss';
 import commonjs from '@rollup/plugin-commonjs';
 import alias from '@rollup/plugin-alias';
 import json from '@rollup/plugin-json';
-import yaml from '@rollup/plugin-yaml';
+// import yaml from '@rollup/plugin-yaml';
 import resolve from '@rollup/plugin-node-resolve';
 // import typescript from '@rollup/plugin-typescript';
+import dynamicImportVars from '@rollup/plugin-dynamic-import-vars';
 import sucrase from '@rollup/plugin-sucrase';
-import {terser} from 'rollup-plugin-terser';
+import terser from '@rollup/plugin-terser';
+
+// import {terser} from 'rollup-plugin-terser';
 
 import nodePolyfills from 'rollup-plugin-polyfill-node';
-import dynamicImportVars from '@rollup/plugin-dynamic-import-vars';
+
+import yaml from '../../tasks/plugin/rollup-plugin-yaml.mjs';
 
 
 let drop_console = false;